fix(server): address review findings on the sqlx Any Postgres port
- Bound concurrent HTTP imports with a semaphore (2) so imports cannot
  drain the 4-connection pool, and run the import pipeline on blocking
  threads via spawn_blocking again
- Dialect-aware BEGIN IMMEDIATE for SQLite staging, promote, and guest
  clone transactions (plain BEGIN on Postgres) so concurrent writers
  cannot race into SQLITE_BUSY
- Case-insensitive LIKE through the like_ci dialect helper for the
  to:/from:/with: filters so both engines return the same results
- Postgres prefix search falls back to plainto_tsquery for terms a
  tsquery literal cannot carry (quotes/backslashes) instead of 500ing
- sqlite:// URL pools now create missing files and enable WAL,
  matching open_pool_for_path
- Serialize the one-time Postgres DDL install with pg_advisory_xact_lock
  and gate it behind a schema_meta marker (vault_schema_v1) so requests
  no longer re-run the full DDL batch
- Demo-guest assignment takes FOR UPDATE on Postgres so concurrent
  try-demo requests cannot double-assign a ready guest
- Refuse reset-demo when the config serves a database URL (the reset
  would only touch a file the vault is not using)
- Honor --db-url for account resolution in import and
  dedupe-cross-source
- Widen temp-table id columns and the guest-expiry cast to BIGINT for
  the Postgres port
- Cross-process PG test lock (fs2 file lock plus the in-process mutex)
  so the lib and integration test binaries cannot race on the shared
  test database
- Refresh the stale Cargo.lock to match the manifests

+++ crates/vault/server/src/dedupe.rs
@@ -373,7 +373,7 @@ async fn recompute_content_keys(
     for stmt in schema::split_ddl(
         r#"
         CREATE TEMP TABLE IF NOT EXISTS _content_keys (
-            id INTEGER PRIMARY KEY,
+            id BIGINT PRIMARY KEY,
             content_key TEXT NOT NULL
         );
         DELETE FROM _content_keys;
@@ -743,8 +743,8 @@ async fn apply_duplicate_flags(
 ) -> Result<()> {
     for stmt in schema::split_ddl(&format!(
         "CREATE TEMP TABLE IF NOT EXISTS {table} (
-            id INTEGER PRIMARY KEY,
-            winner INTEGER NOT NULL
+            id BIGINT PRIMARY KEY,
+            winner BIGINT NOT NULL
         );
         DELETE FROM {table};"
     )) {
